Fixes postgres port default value (#615)
* Fixes postgres port default value

Updates the default port value for the PostgreSQL instance
to be a string, ensuring compatibility with the configuration schema.

Updates the requirements file to use a specific version of
quixstreams[postgresql] instead of a direct github link.

* Updates quixstreams dependencies to 3.14.1.

Updates quixstreams dependencies in destination connectors to version 3.14.1.
Removes git+ install method in favor of standard versioning.