Enables authentication for MQTT connections (#622)
* Enables authentication for MQTT connections

Adds support for username/password authentication to the MQTT client.

The client now checks for the presence of `mqtt_username` environment variable. If it exists, TLS is enabled and the authentication is configured accordingly. This ensures that authentication is only enabled when credentials are provided.

* Refactors MQTT authentication configuration

Changes the authentication configuration to accept the username as a parameter.
This modification allows for greater flexibility in how usernames are handled,
particularly when usernames are dynamically determined or retrieved from
sources other than environment variables. It enforces that a password must
be set if a username is provided.

* Configures MQTT authentication with username

Updates the MQTT client configuration to pass the username
to the authentication function. This allows for
authenticated connections to the MQTT broker.
Adds a message when anonymous authentication is used.