Release main to release/v3 (#632)
* Update PostgreSQLSink template to handle new Primary Keys update (#629)

* Replace hidden inputs with secrets + minor MQTT improvements (#631)

* Updates input type to Secret for sensitive fields

Updates the InputType from HiddenText to Secret for fields containing sensitive information like passwords, API keys and secrets across multiple destinations and sources.
This ensures proper handling and storage of credentials within the system.

* Adds default value for MQTT topic

Adds a default value for the MQTT topic environment variable
to ensure the integration functions correctly out-of-the-box.

* Uses deployment ID for client ID

Changes the environment variable used to generate the MQTT client ID from `Quix__Deployment__Name` to `Quix__Deployment__Id`.

This ensures that the client ID is unique and stable across deployments, as the deployment name may not always be unique.

* Refactors Kafka key to topic name

Updates the Kafka producer key to use the MQTT topic name instead of a modified version.

This change improves clarity and consistency in message routing by ensuring the Kafka key directly reflects the source MQTT topic.

---------

Co-authored-by: Tim Sawicki <136370015+tim-quix@users.noreply.github.com>