Marimo template (#661)
* Adds Marimo notebook template

Introduces a new template for running Marimo reactive notebooks within the platform.

This template includes a Dockerfile for containerization, a Marimo application definition, and necessary configurations for deployment and environment variables (e.g., password protection, QuixLake URL).

* Simplifies Marimo app setup

Streamlines initial configuration by removing environment variable dependencies and providing in-code defaults for the QuixLake URL and SQL query. This makes it easier for new users to get started with the template app.

* Removes marimo app configuration file

Removes the marimo app configuration file, as it's no longer required.

* Configures application directory and path

Adds an argument to specify the main application path
within the Docker image. This enhances flexibility and
organization in deploying marimo applications.

* Updates marimo dockerfile for quixlake sdk

Updates the marimo dockerfile to use the new version of quixlake sdk.
It also simplifies the dockerfile by removing the MAINAPPPATH argument.

* Adds MARIMO_MODE environment variable

Introduces MARIMO_MODE environment variable to control Marimo's execution mode (run/edit).

This variable allows users to configure whether Marimo runs in read-only (production) mode or in editable (development) mode. It defaults to 'edit'.

* Adds support for customizable notebook file

Introduces `MARIMO_NOTEBOOK` environment variable to allow
users to specify the notebook file to open in the marimo
application. This provides more flexibility in deploying
marimo applications with different notebook configurations.

* Fetches quixlake SDK from GitHub

Updates the Dockerfile to download and install the quixlake SDK directly from a GitHub URL, ensuring the correct version is used and simplifying the build process.

Removes the need to manually copy the wheel file into the container.

* Installs quixlake SDK from GitHub URL directly

Simplifies the Dockerfile by installing the quixlake SDK directly from its GitHub URL, removing the need to download and then install the file. This streamlines the build process and reduces the number of steps.

* Adds browser mode to the marimo app

Adds a "browser" mode to the marimo application, allowing users to navigate the file system through a web interface.
This is achieved by running marimo in "edit" mode without a specified notebook.
Updates the documentation to reflect the new mode.

* Updates quixlake_sdk wheel file.

Updates the quixlake_sdk wheel file to the latest version.

This ensures that the project uses the most recent features and
bug fixes provided by the library.

* Updates Quixlake SDK to version 0.2.1

Updates the Quixlake SDK to the latest version (0.2.1) to incorporate the newest features and bug fixes, ensuring compatibility and improved functionality for the marimo template.